requirements.txt
numpy==1.20.3
drawSvg==1.8.1
Pillow==8.2.0